move pkg/controllermanager to cmd/kube-controller-manager/app
