+++ cmd/kube-controller-manager/controller-manager.go
@@ -25,7 +25,7 @@ import (
 	"os"
 	"runtime"
 
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/controllermanager"
+	"github.com/GoogleCloudPlatform/kubernetes/cmd/kube-controller-manager/app"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/util"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/version/verflag"
 
@@ -34,7 +34,7 @@ import (
 
 func main() {
 	runtime.GOMAXPROCS(runtime.NumCPU())
-	s := controllermanager.NewCMServer()
+	s := app.NewCMServer()
 	s.AddFlags(pflag.CommandLine)
 
 	util.InitFlags()
